Combining the EncryptionUtils and refactoring 2 tests.
